+++ src/components/Admin/AccountManagement/ChangePassword.tsx
@@ -18,7 +18,7 @@ import {
   handleReauthenticationWithPassword,
   PasswordFormInput,
   SubmitButton,
-} from './Util';
+} from '../Util';
 import {firebaseAuth} from '../../../firebase';
 import {useTranslation} from 'react-i18next';
